requirements.txt
# Sound2Sheet Requirements

# Basic dependencies for current development phase
numpy>=1.24.0
pyyaml>=6.0

# Audio processing (will be added when needed)
# librosa>=0.10.0
# soundfile>=0.12.0

# Development tools
# black>=23.0.0
# isort>=5.12.0
# flake8>=6.0.0

# Note: Additional dependencies will be added as features are implemented